jit: Use "uint8_t" for ->reg in struct live_interval
This patch shrinks struct live_interval from 76 bytes to 72 bytes on 32-bit by
using "uint8_t" instead of "enum machine_reg" which is 4 bytes long.

+++ jit/linear-scan.c
@@ -164,7 +164,7 @@ static void __spill_interval_intersecting(struct live_interval *current,
 {
 	unsigned long start;
 
-	if (it->reg != reg)
+	if (it->reg != (int8_t) reg)
 		return;
 
 	if (!intervals_intersect(it, current))
